+++ packages/components/CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Checkbox`: supports now being wrapped inside a native `label` element to provide more flexibility in positioning a
   custom label ([#4112](https://github.com/porsche-design-system/porsche-design-system/pull/4112))
+- `Checkbox`, `Radio Button Group`: New slot `end` to add custom content (e.g. `Popover`) at the end of the `label`
+  element ([#4112](https://github.com/porsche-design-system/porsche-design-system/pull/4112))
 - `Multi Select, Select`:
   - `filter` slot to allow custom asynchronous filtering
   - `toggle` event when opening/closing the dropdown